requirement.txt
easydict==1.9
matplotlib==3.0.0
mxnet-cu90==1.3.1b20181119
numba==0.40.1
numpy==1.14.5
Pillow==5.3.0
PyYAML==3.13
requests==2.20.1
scikit-image==0.14.1
scikit-learn==0.20.0
scipy==1.1.0
tqdm==4.28.1

